+++ javaconcepts/src/main/java/com/javaconcepts/app/App.java
@@ -1,0 +1,16 @@
+package com.javaconcepts.app;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main(String[] args)
+    {
+        App2 newapp = new App2();
+        System.out.println("Hello World!");
+        System.out.println("Welcome to my first App");
+        newapp.displayName();        
+    }
+}
