+++ javaconcepts/src/main/java/com/javaconcepts/app/App2.java
@@ -1,0 +1,9 @@
+package com.javaconcepts.app;
+
+public class App2 
+{
+    public void displayName()
+    {
+        System.out.println("Welcome to my second class");
+    }
+}
