--- pom.xml
@@ -1,119 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>com.tutorial.app</groupId>
-  <artifactId>tutorialapp</artifactId>
-  <version>1.0-SNAPSHOT</version>
-
-  <name>tutorialapp</name>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
-    <sonar.host.url>http://192.168.1.130:9000</sonar.host.url>
-    <sonar.login>77735d7e883229a4b277936dddfa9871a794bb86</sonar.login>
-    <sonar.projectName>tutorialapp</sonar.projectName>
-    <sonar.projectVersion>master</sonar.projectVersion>
-    <sonar.test.exclusions>**/test/**/*.*</sonar.test.exclusions>
-    <sonar.exclusions>**/ai/**/*.*,**/jdbc/**/*.*,**/mpt/**/*.*,**/jcr/**/*.*,**/JDBC*</sonar.exclusions>
-
-    <sonar.java.coveragePlugin>jacoco</sonar.java.coveragePlugin>
-    <sonar.dynamicAnalysis>reuseReports</sonar.dynamicAnalysis>
-    <sonar.jacoco.reportPath>target/jacoco.exec</sonar.jacoco.reportPath>
-    <sonar.coverage.jacoco.xmlReportPaths>target/site/jacoco/jacoco.xml</sonar.coverage.jacoco.xmlReportPaths>
-    <sonar.language>java</sonar.language>
-    <jacoco.version>0.8.4</jacoco.version>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
-      <plugins>
-        <!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
-        <plugin>
-          <artifactId>maven-clean-plugin</artifactId>
-          <version>3.1.0</version>
-        </plugin>
-        <!-- default lifecycle, jar packaging: see https://maven.apache.org/ref/current/maven-core/default-bindings.html#Plugin_bindings_for_jar_packaging -->
-        <plugin>
-          <artifactId>maven-resources-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <groupId>org.sonarsource.scanner.maven</groupId>
-          <artifactId>sonar-maven-plugin</artifactId>
-          <version>3.7.0.1746</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.8.0</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-surefire-plugin</artifactId>
-          <version>2.22.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-jar-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-install-plugin</artifactId>
-          <version>2.5.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-deploy-plugin</artifactId>
-          <version>2.8.2</version>
-        </plugin>
-        <!-- site lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#site_Lifecycle -->
-        <plugin>
-          <artifactId>maven-site-plugin</artifactId>
-          <version>3.7.1</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-project-info-reports-plugin</artifactId>
-          <version>3.0.0</version>
-        </plugin>
-        <plugin>
-	    <groupId>org.jacoco</groupId>
-	    <artifactId>jacoco-maven-plugin</artifactId>
-	    <version>${jacoco.version}</version>
-	    <configuration>
-		<skip>${maven.test.skip}</skip>
-		<destFile>${basedir}/target/coverage-reports/jacoco-unit.exec</destFile>
-		<dataFile>${basedir}/target/coverage-reports/jacoco-unit.exec</dataFile>
-		<output>file</output>
-		<append>true</append>
-		<excludes>
-		    <exclude>*MethodAccess</exclude>
-		</excludes>
-	    </configuration>
-	    <executions>
-		<execution>
-		    <goals>
-			<goal>prepare-agent</goal>
-		    </goals>
-		</execution>
-		<execution>
-		    <id>report</id>
-                    <phase>prepare-package</phase>  
-		    <goals>
-			<goal>report</goal>
-		    </goals>
-		</execution>
-	    </executions>
-        </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
-</project>
